batchconvert: Run video encoder in seperate thread
x265 doesn't close properly when CTRL-C is used, and can only be closed
via SIGTERM by a seperate process (such as pkill).

Since x265 takes the foreground of the main python thread when it's running python
is unable to react to KeyboardInterrupt until x265 closes. By running it
in a seperate python thread the main thread can react to
KeyboardInterrupt and send a SIGTERM to x265.