batchconvert: Custom 'encodeInfo.py's per folder
By placing an 'encodeInfo.py' file inside a folder along with the
'info.json' file the script will use that 'encodeInfo.py' file instead
of the one next to the 'batchconvert.py' script.

This is for videos with different resolutions that may require a
different set of vapoursynth filters. Like bonus features where some are
1080p, but others are 480p with interlacing.