nightmode/create-nightmode: Don't show duration if we can't fine one
Some codecs and containers don't provide the metadata that show the
duration of the audio track. (mainly raw formats like TrueHD or DTSHD)

In my script it would simply stop running if it can't display the
file's duration. Since the duration is not nessesary for the script to
work we'll just ignore it if we can't find it.