video-hdr: Add option to skip re-encode
Sometimes I'm too impatient to wait for a re-encode, and I just want to
watch on my setup (which is able to handle most 4K bluray remuxes).

This will just extract the video track without re-encoding it with x265.

If Dolby Vision is detected it'll convert the RPU to Profile 8.1 and
inject it into the source video track via 'dovi_tool'.