batchconvert: Handle HDR content
Finally migrate code in 'video-hdr.py' into batchconvert. This allows
batchconvert to detect and handle HDR10, HDR10+, and Dolby Vision video
automatically.

The only thing not implemented yet is handling Dolby Vision when
'convert' is set to 'false'. Dolby Vision needs to be converted to
Profile 8.1 in order for the video to be compatible with most devices.